backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-binance
ccxt
pandas
numpy
statsmodels
scipy
aiohttp
python-dotenv
